add switch to share
