+++ package.json
@@ -26,6 +26,7 @@
     "react-outside-click-handler": "^1.3.0",
     "react-slick": "^0.25.2",
     "react-tippy": "^1.3.4",
+    "react-toggle": "^4.1.1",
     "release": "^6.1.0"
   },
   "devDependencies": {
