+++ src/components/switch/index.js
@@ -1,3 +1,43 @@
-import Component from './switch-component';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
+import Toggle from 'react-toggle';
 
-export default Component;
+import './react-toggle.scss';
+import './styles.scss';
+import './themes/light.scss';
+
+class Switch extends PureComponent {
+  static propTypes = {
+    theme: PropTypes.string,
+    label: PropTypes.string,
+    value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+    options: PropTypes.array,
+    onChange: PropTypes.func,
+    className: PropTypes.string,
+  };
+
+  render() {
+    const { theme, label, value, options, onChange, className } = this.props;
+
+    return (
+      <div className={`c-switch ${theme || ''} ${className || ''}`}>
+        {label && <div className="label">{label}</div>}
+        <Toggle
+          icons={{
+            checked: options[0].label,
+            unchecked: options[1].label,
+          }}
+          defaultChecked={options[1].value === value}
+          onChange={(e) => {
+            const result = e.target.checked
+              ? options[1].value
+              : options[0].value;
+            onChange(result);
+          }}
+        />
+      </div>
+    );
+  }
+}
+
+export default Switch;
