+++ src/components/switch/styles.scss
@@ -2,7 +2,7 @@
 
 .c-switch {
   .label {
-    margin-bottom: rem(5px);
+    margin-bottom: 5px;
     display: flex;
     justify-content: flex-start;
     align-items: center;
