+++ src/components/switch/themes/light.scss
@@ -7,7 +7,7 @@
 
   .react-toggle-track {
     width: 100%;
-    height: rem(34px);
+    height: 34px;
     background-color: transparent !important;
     border: solid 1px $border;
 
@@ -37,7 +37,7 @@
     height: 100%;
     background-color: transparent;
     border-radius: 30px;
-    border: solid 1px $green-gfw;
+    border: solid 1px $green;
     box-shadow: none !important;
   }
